[feature] Migrate the main.go directory
