+++ conf/koala.conf
@@ -26,10 +26,10 @@ externalWriteTimeout = 500
 #--------------
 
 #redis服务地址
-redis_server = 127.0.0.1:3721
+redis_server = 127.0.0.1:6379
 
 #redis服务password
-redis_auth = Yhs8jCfcys
+redis_auth =
 
 #redis连接池，最大空闲连接数
 redis_pool_maxIdle = 5
@@ -52,10 +52,10 @@ log_warning_file_path	= log/koala.log.wf
 log_cron_time = day
 
 #日志文件生存周期, 单位:天
-log_file_ilfe_time = 7
+log_file_life_time = 7
 
-#日志chan队列的buffer长度，建议不要少于10240，不建议多于1024000，最长：67021478（超过这个值会无法启动）
-log_chan_buff_size = 10240
+#日志channel队列的buffer长度，建议不要少于10240，不建议多于1024000，最长：67021478（超过这个值会无法启动）
+log_channel_buff_size = 10240
 
 #日志刷盘的时间间隔，单位：毫秒，建议500~5000毫秒，建议不超过30秒
 log_flush_timer = 1000
