+++ go.mod
@@ -2,4 +2,7 @@ module github.com/heiyeluren/koala
 
 go 1.16
 
-require github.com/gomodule/redigo v1.8.5
+require (
+	github.com/gomodule/redigo v1.8.5
+	gopkg.in/yaml.v2 v2.3.0 // indirect
+)
